+++ tests/integration/test_async_http_client.py
@@ -1,5 +1,9 @@
 """Tests for the HTTP API Client."""
+from typing import Tuple
 import pytest
+
+from solders.rpc.requests import GetBlockHeight, GetFirstAvailableBlock
+from solders.rpc.responses import GetBlockHeightResp, GetFirstAvailableBlockResp, Resp
 
 import solana.system_program as sp
 from solana.blockhash import Blockhash
@@ -521,3 +525,17 @@ async def test_get_vote_accounts(test_http_client_async):
     """Test get vote accounts."""
     resp = await test_http_client_async.get_vote_accounts()
     assert_valid_response(resp)
+
+
+@pytest.mark.integration
+async def test_batch_request(test_http_client_async: AsyncClient):
+    """Test get vote accounts."""
+    reqs = (GetBlockHeight(), GetFirstAvailableBlock())
+    parsers = (GetBlockHeightResp, GetFirstAvailableBlockResp)
+    resp: Tuple[
+        Resp[GetBlockHeightResp], Resp[GetFirstAvailableBlockResp]
+    ] = await test_http_client_async._provider.make_batch_request(  # pylint: disable=protected-access
+        reqs, parsers
+    )
+    assert_valid_response(resp[0])
+    assert_valid_response(resp[1])